Changed up How to delete items
Changed to way to delete book items. Instead of a user using a input field to delete a string, now bound icons to the books so that you can simply click delete on the item you want. I don't like that I had to add specific margin on the container for the delete icons so that they would line up with the list items.

Also, added in the use of FontAwesome in order to get the desired icon.